This is a temporary runner
I'm using this to test my code because the runner you wrote eddie isn't fully functional. I have wrote a lot more code, but will not be putting it out because I started the entire project from scratch and i dont want to erase anything. I've reworked the deck class so that the scanner is in the runner and not in the actual class, and have done this via the above way. I'll load the rest of what I've done when I think that it's good enough.